Trait kff::data::Reader

pub trait Reader<'input, R>: Iterator<Item = Result<Kmer>> where
    R: Read + 'input,
    R: ?Sized, {
    fn k(&self) -> u64;

    fn input(&mut self) -> &mut R;

    fn max_kmer(&self) -> u64;

    fn block_n(&self) -> u64;

    fn data_size(&self) -> u64;

    fn read_block(&mut self) -> Result<usize>;

    fn block_seq(&self) -> &Seq2Slice;

    fn block_data(&self) -> &[u8];

    fn rev_encoding(&self) -> u8;

    fn decrease_n(&mut self);

    fn read_n(&mut self) -> Result<u64> { ... }

    fn read_seq(&mut self, nb_nuc: u64) -> Result<BitBox> { ... }

    fn read_data(&mut self) -> Result<Vec<u8>> { ... }

    fn next_kmer(&mut self) -> Result<Kmer> { ... }
}

Reader trait must be implement by Struct want read kmer KFF section

Required methods

fn k(&self) -> u64

Get size of kmers

fn input(&mut self) -> &mut R

Get a reference to the read stream

fn max_kmer(&self) -> u64

Get the max number of kmer in a block

fn block_n(&self) -> u64

Get the actual number of kmer remains in block

fn data_size(&self) -> u64

Get the size of data attach to each kmer in bytes

fn read_block(&mut self) -> Result<usize>

Read on kmer block and return number of bytes read

fn block_seq(&self) -> &Seq2Slice

Get the sequence of actual block

fn block_data(&self) -> &[u8]

Get the data of actutal block

fn rev_encoding(&self) -> u8

Get bit used to perform reverse encoding

fn decrease_n(&mut self)

Reduce block_n by one

Loading content...
Provided methods

fn read_n(&mut self) -> Result<u64>

Read the number of kmer in block

fn read_seq(&mut self, nb_nuc: u64) -> Result<BitBox>

Read sequence of actual block need number of nucleotide want be to read

fn read_data(&mut self) -> Result<Vec<u8>>

Read data of actual block

fn next_kmer(&mut self) -> Result<Kmer>

Get the next kmer

Loading content...