Trait kff::data::Writer

pub trait Writer<'output, W>: Sized where
    W: Write + Seek + 'output, {
    fn data_size(&self) -> u64;

    fn max(&self) -> u64;

    fn nb_block_offset(&self) -> u64;

    fn nb_block(&self) -> u32;

    fn is_close(&self) -> bool;

    fn output(&mut self) -> &mut W;

    fn set_nb_block(&mut self, value: u32);

    fn set_close(&mut self, value: bool);

    fn nb_kmer(&self, seq_len: usize) -> usize;

    fn close(&mut self) -> Result<usize> { ... }

    fn increment_nb_block(&mut self) -> Result<()> { ... }

    fn check_block(&mut self, seq_len: usize, data_len: usize) -> Result<usize> { ... }
}

Write trait must be implement by Struct want write kmer KFF section

Required methods

fn data_size(&self) -> u64

Get the size of data attach to each kmer in bytes

fn max(&self) -> u64

Get the max number of kmer in a block

fn nb_block_offset(&self) -> u64

Position of nb_block_offset need to write the number of block in section after finish to write section

fn nb_block(&self) -> u32

Number of block in this section

fn is_close(&self) -> bool

Return True if section is close don't write any other block in this section

fn output(&mut self) -> &mut W

Get a reference of output

fn set_nb_block(&mut self, value: u32)

Change number of block in section

fn set_close(&mut self, value: bool)

Change state of Writer::is_close

fn nb_kmer(&self, seq_len: usize) -> usize

Compute the number of kmer from number of sequence

Loading content...
Provided methods

fn close(&mut self) -> Result<usize>

Close the section, can return an error because number of block can only by write when number of block is know

fn increment_nb_block(&mut self) -> Result<()>

Increase the number of block

fn check_block(&mut self, seq_len: usize, data_len: usize) -> Result<usize>

Verify the number of the number of kmer and data length match and not upper than max number of kmer

Loading content...