Struct kff::kff::Kff

pub struct Kff<T> { /* private fields */ }

Struct to read a kff file

Implementations

impl<T> Kff<T>

pub fn inner(&self) -> &T

Inner read source

pub fn header(&self) -> &Header

Header extract from inner

pub fn values(&self) -> &Values

Current Values extract from inner

pub fn index(&self) -> &Option<GlobalIndex>

GlobalIndex present only if inner is seekable and first section is index or footer contains first_index

impl<T> Kff<T>

pub fn set_header(&mut self, val: Header) -> &mut Self

Header extract from inner

pub fn set_values(&mut self, val: Values) -> &mut Self

Current Values extract from inner

impl<T> Kff<T>

pub fn header_mut(&mut self) -> &mut Header

Header extract from inner

pub fn values_mut(&mut self) -> &mut Values

Current Values extract from inner

impl<R> Kff<R>
where R: Read + BufRead + KffRead,

pub fn read(inner: R) -> Result<Self>

Create a new Kff reader by accept mutable reference on std::io::Read

pub fn kmers(self) -> KmerIterator<R>

Consume Kff object to create a KmerIterator

pub fn next_kmer_section(&mut self) -> Option<Result<Vec<Kmer>>>

Read Kff until last kmer section

impl Kff<BufReader<File>>

pub fn open<P>(path: P) -> Result<Self>
where P: AsRef<Path>,

Create a new Kff by read file match with path

pub fn with_index<P>(path: P) -> Result<Self>
where P: AsRef<Path>,

Create a Kff and generate a global index

impl<R> Kff<R>
where R: Read + Seek + KffRead,

pub fn check(&mut self) -> Result<bool>

Check readable match with a KFF file

pub fn kmer_of_section(&mut self, n: usize) -> Result<Vec<Kmer>>

Get kmer of nth section in index.

If index isnโt set return an Error If we didnโt found section value before target section return an Error If section isnโt a kmer section return an Error

impl<W> Kff<W>
where W: Write + KffWrite,

pub fn write(inner: W, header: Header) -> Result<Self>

Create a Kff object to write in inner

pub fn write_values(&mut self, values: Values) -> Result<()>

Write a Values section

pub fn write_index(&mut self, index: Index) -> Result<()>

Write a Index section

pub fn write_raw(&mut self, raw: Raw, blocks: &[Block]) -> Result<()>

Write a Raw section

pub fn write_minimizer( &mut self, section: Minimizer, minimizer: Seq2Bit, blocks: &[Block], ) -> Result<()>

Write a Minimizer section

pub fn finalize(&mut self) -> Result<()>

Finalize write the final signature

impl Kff<BufWriter<File>>

pub fn create<P>(path: P, header: Header) -> Result<Self>
where P: AsRef<Path>,

Intialize a Kff object to write file