Trait kff::section::values::AbcValues

pub trait AbcValues: Sized {
    fn new() -> Self;
    fn with_capacity(capacity: usize) -> Self;
    fn read<R>(inner: &mut R) -> Result<Self>
    where
        R: Read + KffRead;
    fn write<W>(&self, outer: &mut W) -> Result<()>
    where
        W: Write;
    fn write_as_footer<W>(&self, outer: &mut W) -> Result<()>
    where
        W: Write + KffWrite;
}

A trait to implement some function around Values alias

Required Methods

fn new() -> Self

Build an empty Values

fn with_capacity(capacity: usize) -> Self

Build an empty Values but with a minimal capacity

fn read<R>(inner: &mut R) -> Result<Self> where
R: Read + KffRead,

Build a Values from a readable key are overwrite, section flag must be already read

fn write<W>(&self, outer: &mut W) -> Result<()> where
W: Write,

Write contents of Values in writables, section flag isn’t write

fn write_as_footer<W>(&self, outer: &mut W) -> Result<()> where
W: Write + KffWrite,

Write contents of Values in a writables as a footer